Ok
I'm going to be pretty short on time soon. But I will try to improve or add to the translation as it comes. 
At least as soon as a major update is released I will have to revise the translation of the new level tree. Right now I think I've translated everything I can.